Add Test norepeat strategy
Wall-E should always repeat itself when confronted to successive
occurrences of certain commands (like "help").

Wall-E should not repeat itself when confronted twice in a row with the same
blocking error.

Wall-E should repeat the message for error A when confronted to error A, then
error B, then error A again.