Wait a few seconds before deleting repository
fixes: sometimes the test repository created for wall-e tests cannot
be deleted anymore. The content is deleted, but the repo stays visible
on Bitbucket forever, and it is not possible to delete it from the
interface, neither from the API. I suspect this happens mostly when
tests are run individually, and we request the repo to be deleted
very shortly after it's creation. As a workaround, wait a few seconds
before deleting the repository.

RELENG-1542