RELENG-2741 Document the permanent bypass of Fix Version checks
The previous commits have added the capability to disable the version Checks in
Bert-E gating logic. The documentation was not updated to reflect this.

Although Bert-E does not have a lot of documentation, it seemed relevant to
document in the USER Doc that some behaviors could be deactivated through the
configuration.