BERTE-487: Ignore github COMMENTED reviews for approval accounting
The previous logic from the github githost implementation for the
`get_approvals()` function was taking into account the last review (from the
API) status.

As that status could be any of ("APPROVED", "COMMENTED", "DISMISSED",
"CHANGES_REQUESTED"), we needed to account only for the relevant statuses to be
aligned with the github UI's behavior. Indeed, "COMMENTED" statuses have never
been taken into account by github as contradiction to an approval, while a
request change is clearly shown in red (and as such, as a "bad" state).