BERTE-487: githosts: Add CHANGES_REQUESTED support
As observed during the investigation for BERTE-487, CHANGES_REQUESTED review
status usually means from github's UI point-of-view that the previous approvals
are no longer accounted for (the summary at the bottom of the PR no longer
shows approved reviews afterwards, despite the approved review showing in the
PR's discussion history).
Additionally, "DISMISSED" status actually overrides the status affected
(usually the last "APPROVED" or "CHANGES_REQUESTED" from the reviewer), so it
is not possible to know after the facts whether the dismissed review was an
approval or a request for changes.

As a whole, this means that the three statuses can be seen as exclusive for any
reviewer, and only the last one should be accounted for, which aligns Bert-e's
behavior with github's UI's.