BERTE-487: Github: Add review dismissal support
The previous logic from the github githost implementation for the
`get_approvals()` function was taking into account the last review (from the
API) status.

This commit provides the capability to dismiss a review (no matter its state),
and comes with a test validated against github. This shall be used later to
validate a full understanding of the github review states, through additional
tests.