BERTE-502 renaming Bert-E robot username config
In order to be more coherent with the incoming changes in Bert-E's
code. We're renaming the robot_username settings to robot.

Following usage of the robot setting will be like:
```python
robot.username = toto
robot.account_id = 42
```

therefore it doesn't make sense anymore to keep the name of
`robot_username`